pyqcrbox/requirements-docs.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile --extra=docs --output-file=requirements-docs.txt pyproject.toml
annotated-types==0.7.0
    # via pydantic
anyio==4.4.0
    # via
    #   fast-depends
    #   faststream
    #   httpx
appnope==0.1.4
    # via ipykernel
asttokens==2.4.1
    # via stack-data
attrs==23.2.0
    # via
    #   jsonschema
    #   referencing
    #   svcs
babel==2.15.0
    # via mkdocs-material
beautifulsoup4==4.12.3
    # via nbconvert
bleach==6.1.0
    # via nbconvert
bracex==2.4
    # via wcmatch
certifi==2024.7.4
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   hatch
    #   mkdocs
    #   typer
    #   userpath
cloudpickle==3.0.0
    # via doit
colorama==0.4.6
    # via mkdocs-material
comm==0.2.2
    # via ipykernel
debugpy==1.8.2
    # via ipykernel
decorator==5.1.1
    # via ipython
defusedxml==0.7.1
    # via nbconvert
distlib==0.3.8
    # via virtualenv
doit==0.36.0
    # via pyqcrbox (pyproject.toml)
executing==2.0.1
    # via stack-data
fast-depends==2.4.6
    # via faststream
fastjsonschema==2.20.0
    # via nbformat
faststream==0.5.15
    # via pyqcrbox (pyproject.toml)
filelock==3.15.4
    # via virtualenv
ghp-import==2.1.0
    # via mkdocs
gitdb==4.0.11
    # via gitpython
gitpython==3.1.43
    # via pyqcrbox (pyproject.toml)
greenlet==3.0.3
    # via sqlalchemy
h11==0.14.0
    # via httpcore
hatch==1.12.0
    # via pyqcrbox (pyproject.toml)
hatchling==1.25.0
    # via hatch
httpcore==1.0.5
    # via httpx
httpx==0.27.0
    # via hatch
hyperlink==21.0.0
    # via hatch
idna==3.7
    # via
    #   anyio
    #   httpx
    #   hyperlink
    #   requests
importlib-metadata==8.0.0
    # via doit
ipykernel==6.29.5
    # via mkdocs-jupyter
ipython==8.26.0
    # via ipykernel
jaraco-classes==3.4.0
    # via keyring
jaraco-context==5.3.0
    # via keyring
jaraco-functools==4.0.1
    # via keyring
jedi==0.19.1
    # via ipython
jinja2==3.1.4
    # via
    #   mkdocs
    #   mkdocs-material
    #   nbconvert
jsonschema==4.23.0
    # via nbformat
jsonschema-specifications==2023.12.1
    # via jsonschema
jupyter-client==8.6.2
    # via
    #   ipykernel
    #   nbclient
jupyter-core==5.7.2
    # via
    #   ipykernel
    #   jupyter-client
    #   nbclient
    #   nbconvert
    #   nbformat
jupyterlab-pygments==0.3.0
    # via nbconvert
jupytext==1.16.3
    # via mkdocs-jupyter
keyring==25.2.1
    # via hatch
loguru==0.7.2
    # via pyqcrbox (pyproject.toml)
markdown==3.6
    # via
    #   mkdocs
    #   mkdocs-material
    #   pymdown-extensions
markdown-exec==1.9.3
    # via pyqcrbox (pyproject.toml)
markdown-it-py==3.0.0
    # via
    #   jupytext
    #   mdit-py-plugins
    #   rich
markupsafe==2.1.5
    # via
    #   jinja2
    #   mkdocs
    #   nbconvert
matplotlib-inline==0.1.7
    # via
    #   ipykernel
    #   ipython
mdit-py-plugins==0.4.1
    # via jupytext
mdurl==0.1.2
    # via markdown-it-py
mergedeep==1.3.4
    # via
    #   mkdocs
    #   mkdocs-get-deps
mistune==3.0.2
    # via nbconvert
mkdocs==1.6.0
    # via
    #   pyqcrbox (pyproject.toml)
    #   mkdocs-awesome-pages-plugin
    #   mkdocs-jupyter
    #   mkdocs-material
mkdocs-awesome-pages-plugin==2.9.2
    # via pyqcrbox (pyproject.toml)
mkdocs-get-deps==0.2.0
    # via mkdocs
mkdocs-jupyter==0.24.8
    # via pyqcrbox (pyproject.toml)
mkdocs-material==9.5.29
    # via
    #   pyqcrbox (pyproject.toml)
    #   mkdocs-jupyter
mkdocs-material-extensions==1.3.1
    # via mkdocs-material
more-itertools==10.3.0
    # via
    #   jaraco-classes
    #   jaraco-functools
nats-py==2.8.0
    # via faststream
natsort==8.4.0
    # via mkdocs-awesome-pages-plugin
nbclient==0.10.0
    # via nbconvert
nbconvert==7.16.4
    # via mkdocs-jupyter
nbformat==5.10.4
    # via
    #   jupytext
    #   nbclient
    #   nbconvert
nest-asyncio==1.6.0
    # via ipykernel
packaging==24.1
    # via
    #   hatch
    #   hatchling
    #   ipykernel
    #   jupytext
    #   mkdocs
    #   nbconvert
paginate==0.5.6
    # via mkdocs-material
pandocfilters==1.5.1
    # via nbconvert
parso==0.8.4
    # via jedi
pathspec==0.12.1
    # via
    #   hatchling
    #   mkdocs
pexpect==4.9.0
    # via
    #   hatch
    #   ipython
platformdirs==4.2.2
    # via
    #   hatch
    #   jupyter-core
    #   mkdocs-get-deps
    #   virtualenv
pluggy==1.5.0
    # via hatchling
prompt-toolkit==3.0.47
    # via ipython
psutil==6.0.0
    # via ipykernel
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.2
    # via stack-data
pydantic==2.8.2
    # via
    #   pyqcrbox (pyproject.toml)
    #   fast-depends
    #   pydantic-settings
    #   sqlmodel
pydantic-core==2.20.1
    # via pydantic
pydantic-settings==2.3.4
    # via pyqcrbox (pyproject.toml)
pygments==2.18.0
    # via
    #   ipython
    #   mkdocs-jupyter
    #   mkdocs-material
    #   nbconvert
    #   rich
pymdown-extensions==10.8.1
    # via
    #   pyqcrbox (pyproject.toml)
    #   markdown-exec
    #   mkdocs-material
python-dateutil==2.9.0.post0
    # via
    #   ghp-import
    #   jupyter-client
python-dotenv==1.0.1
    # via pydantic-settings
pyyaml==6.0.1
    # via
    #   jupytext
    #   mkdocs
    #   mkdocs-get-deps
    #   pymdown-extensions
    #   pyyaml-env-tag
pyyaml-env-tag==0.1
    # via mkdocs
pyzmq==26.0.3
    # via
    #   ipykernel
    #   jupyter-client
referencing==0.35.1
    # via
    #   jsonschema
    #   jsonschema-specifications
regex==2024.5.15
    # via mkdocs-material
requests==2.32.3
    # via mkdocs-material
rich==13.7.1
    # via
    #   pyqcrbox (pyproject.toml)
    #   hatch
    #   typer
rpds-py==0.19.0
    # via
    #   jsonschema
    #   referencing
shellingham==1.5.4
    # via
    #   hatch
    #   typer
six==1.16.0
    # via
    #   asttokens
    #   bleach
    #   python-dateutil
smmap==5.0.1
    # via gitdb
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
soupsieve==2.5
    # via beautifulsoup4
sqlalchemy==2.0.31
    # via sqlmodel
sqlmodel==0.0.20
    # via pyqcrbox (pyproject.toml)
stack-data==0.6.3
    # via ipython
stamina==24.2.0
    # via pyqcrbox (pyproject.toml)
structlog==24.4.0
    # via pyqcrbox (pyproject.toml)
svcs==24.1.0
    # via pyqcrbox (pyproject.toml)
tabulate==0.9.0
    # via pyqcrbox (pyproject.toml)
tenacity==8.5.0
    # via stamina
tinycss2==1.3.0
    # via nbconvert
tomli-w==1.0.0
    # via hatch
tomlkit==0.13.0
    # via hatch
tornado==6.4.1
    # via
    #   ipykernel
    #   jupyter-client
traitlets==5.14.3
    # via
    #   comm
    #   ipykernel
    #   ipython
    #   jupyter-client
    #   jupyter-core
    #   matplotlib-inline
    #   nbclient
    #   nbconvert
    #   nbformat
trove-classifiers==2024.7.2
    # via hatchling
typer==0.12.3
    # via faststream
typing-extensions==4.12.2
    # via
    #   faststream
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typer
urllib3==2.2.2
    # via requests
userpath==1.9.2
    # via hatch
uv==0.2.27
    # via hatch
virtualenv==20.26.3
    # via hatch
watchdog==4.0.1
    # via mkdocs
wcmatch==8.5.2
    # via mkdocs-awesome-pages-plugin
wcwidth==0.2.13
    # via prompt-toolkit
webencodings==0.5.1
    # via
    #   bleach
    #   tinycss2
zipp==3.19.2
    # via importlib-metadata
zstandard==0.23.0
    # via hatch
